Move ContentId to System.Reflection.PortableExecutable namespace
